move typed api protection interface into cop
