+++ lib/package_protections/private/configuration.rb
@@ -28,7 +28,7 @@ module PackageProtections
         [
           Private::OutgoingDependencyProtection.new,
           Private::IncomingPrivacyProtection.new,
-          Private::TypedApiProtection.new,
+          RuboCop::Cop::PackageProtections::TypedPublicApi.new,
           RuboCop::Cop::PackageProtections::NamespacedUnderPackageName.new,
           Private::VisibilityProtection.new
         ]
